engine,plan,execution: string and number literals as logical nodes
Convert more PromQL AST Expressions into Thanos logical nodes. This is
an ongoing effort, this PR only does NumberLiteral and StringLiteral.
