+++ execution/function/relabel.go
@@ -13,11 +13,10 @@ import (
 	"github.com/efficientgo/core/errors"
 	prommodel "github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/model/labels"
-
 	"github.com/prometheus/prometheus/promql/parser"
 
 	"github.com/thanos-io/promql-engine/execution/model"
-	"github.com/thanos-io/promql-engine/extexpr"
+	"github.com/thanos-io/promql-engine/logicalplan"
 	"github.com/thanos-io/promql-engine/query"
 )
 
@@ -85,7 +84,7 @@ func (o *relabelOperator) loadSeries(ctx context.Context) (err error) {
 }
 
 func (o *relabelOperator) loadSeriesForLabelJoin(series []labels.Labels) error {
-	labelJoinDst, err := extexpr.UnwrapString(o.funcExpr.Args[1])
+	labelJoinDst, err := logicalplan.UnwrapString(o.funcExpr.Args[1])
 	if err != nil {
 		return errors.Wrap(err, "unable to unwrap string argument")
 	}
@@ -94,12 +93,12 @@ func (o *relabelOperator) loadSeriesForLabelJoin(series []labels.Labels) error {
 	}
 
 	var labelJoinSrcLabels []string
-	labelJoinSep, err := extexpr.UnwrapString(o.funcExpr.Args[2])
+	labelJoinSep, err := logicalplan.UnwrapString(o.funcExpr.Args[2])
 	if err != nil {
 		return errors.Wrap(err, "unable to unwrap string argument")
 	}
 	for j := 3; j < len(o.funcExpr.Args); j++ {
-		srcLabel, err := extexpr.UnwrapString(o.funcExpr.Args[j])
+		srcLabel, err := logicalplan.UnwrapString(o.funcExpr.Args[j])
 		if err != nil {
 			return errors.Wrap(err, "unable to unwrap string argument")
 		}
@@ -123,22 +122,22 @@ func (o *relabelOperator) loadSeriesForLabelJoin(series []labels.Labels) error {
 	return nil
 }
 func (o *relabelOperator) loadSeriesForLabelReplace(series []labels.Labels) error {
-	labelReplaceDst, err := extexpr.UnwrapString(o.funcExpr.Args[1])
+	labelReplaceDst, err := logicalplan.UnwrapString(o.funcExpr.Args[1])
 	if err != nil {
 		return errors.Wrap(err, "unable to unwrap string argument")
 	}
 	if !prommodel.LabelName(labelReplaceDst).IsValid() {
 		return errors.Newf("invalid destination label name in label_replace: %s", labelReplaceDst)
 	}
-	labelReplaceRepl, err := extexpr.UnwrapString(o.funcExpr.Args[2])
+	labelReplaceRepl, err := logicalplan.UnwrapString(o.funcExpr.Args[2])
 	if err != nil {
 		return errors.Wrap(err, "unable to unwrap string argument")
 	}
-	labelReplaceSrc, err := extexpr.UnwrapString(o.funcExpr.Args[3])
+	labelReplaceSrc, err := logicalplan.UnwrapString(o.funcExpr.Args[3])
 	if err != nil {
 		return errors.Wrap(err, "unable to unwrap string argument")
 	}
-	labelReplaceRegexVal, err := extexpr.UnwrapString(o.funcExpr.Args[4])
+	labelReplaceRegexVal, err := logicalplan.UnwrapString(o.funcExpr.Args[4])
 	if err != nil {
 		return errors.Wrap(err, "unable to unwrap string argument")
 	}
