+++ logicalplan/distribute.go
@@ -18,7 +18,6 @@ import (
 	"github.com/prometheus/prometheus/promql/parser/posrange"
 
 	"github.com/thanos-io/promql-engine/api"
-	"github.com/thanos-io/promql-engine/extexpr"
 	"github.com/thanos-io/promql-engine/query"
 )
 
@@ -275,7 +274,7 @@ func (m DistributedExecutionOptimizer) distributeQuery(expr *parser.Expr, engine
 	if allowedStartOffset < startOffset {
 		return *expr
 	}
-	if extexpr.IsConstantExpr(*expr) {
+	if IsConstantExpr(*expr) {
 		return *expr
 	}
 
@@ -468,8 +467,8 @@ func isDistributive(expr *parser.Expr, skipBinaryPushdown bool) bool {
 }
 
 func isBinaryExpressionWithOneConstantSide(expr *parser.BinaryExpr) bool {
-	lhsConstant := extexpr.IsConstantExpr(expr.LHS)
-	rhsConstant := extexpr.IsConstantExpr(expr.RHS)
+	lhsConstant := IsConstantExpr(expr.LHS)
+	rhsConstant := IsConstantExpr(expr.RHS)
 	return (lhsConstant || rhsConstant)
 }
 
@@ -531,7 +530,7 @@ func rewritesEngineLabels(e parser.Expr, engineLabels map[string]struct{}) bool 
 		if !ok || call.Func.Name != "label_replace" {
 			return false
 		}
-		targetLabel := extexpr.UnsafeUnwrapString(call.Args[1])
+		targetLabel := UnsafeUnwrapString(call.Args[1])
 		if _, ok := engineLabels[targetLabel]; ok {
 			result = true
 			return true
